server/ai_core/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
google-generativeai
google-cloud-aiplatform
mediapipe
opencv-python-headless
numpy
pymongo
httpx
pandas 
scikit-learn